Separate tests ### What's done: * Separate tests
